Bugfix: Refactor ApplyFuncOnHunts() to not use List()
This is very slow an leads to a performance bottleneck
